refactor: policy controller package
